Update build.yml
Updated cron schedule to build every Sunday at midnight

Previous cron schedule was meant to run on the first of every month, but was actually set to run on the first of every month if-and-only-if it was a Sunday.

Also use a queue-time Build.OfficialRelease value instead of a variable defined in the build.yml file.